+++ src/main/java/com/tagnumelite/chickens/api/utils/constants/TranslationConstants.java
@@ -4,9 +4,9 @@ public class TranslationConstants {
     public static final String ANALYZER_TOOLTIP_1 = "item.chickens.analyzer.tooltip1";
     public static final String ANALYZER_TOOLTIP_2 = "item.chickens.analyzer.tooltip2";
     public static final String HENHOUSE_TOOLTIP = "henhouse.tooltip";
-    public static final String JEI_THROWING = "gui.chickens.jei.throwing";
-    public static final String JEI_BREEDING = "gui.chickens.jei.breeding";
-    public static final String JEI_LAYING = "gui.chickens.jei.breeding";
+    public static final String JEI_THROWING = "jei.throwing";
+    public static final String JEI_BREEDING = "jei.breeding";
+    public static final String JEI_LAYING = "jei.laying";
     public static final String FLUID_EGG_TOOLTIP = "item.chickens.fluid_egg.tooltip"; //It's like one-off bucket.
     public static final String CHICKEN_LAY_PROGRESS = "entity.chickens.chicken.lay_progress";
     public static final String CHICKEN_NEXT_EGG_SOON = "entity.chickens.chicken.next_egg_soon";
@@ -23,4 +23,5 @@ public class TranslationConstants {
     public static final String CONFIG_DROP_CHANCE_EGG = "config.chickens.drop_chance_egg";
     public static final String CONFIG_DROP_CHANCE_ITEM = "config.chickens.drop_chance_item";
     public static final String NAME = "gui.chickens.name";
+    public static final String NULL_CHICKEN = "entity.chickens.chicken.null";
 }