+++ src/main/java/com/tagnumelite/chickens/client/screens/CoopScreen.java
@@ -3,7 +3,7 @@ package com.tagnumelite.chickens.client.screens;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.tagnumelite.chickens.api.utils.Utils;
-import com.tagnumelite.chickens.client.menus.HenhouseMenu;
+import com.tagnumelite.chickens.client.menus.CoopMenu;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.network.chat.Component;
@@ -11,11 +11,21 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
 import org.jetbrains.annotations.NotNull;
 
-public class HenhouseScreen extends AbstractContainerScreen<HenhouseMenu> {
-    static final ResourceLocation BACKGROUND = Utils.rl("textures/gui/container/henhouse.png");
+public class CoopScreen extends AbstractContainerScreen<CoopMenu> {
+    static final ResourceLocation BACKGROUND = Utils.rl("textures/gui/container/coop.png");
 
-    public HenhouseScreen(HenhouseMenu pMenu, Inventory pPlayerInventory, Component pTitle) {
+    public CoopScreen(CoopMenu pMenu, Inventory pPlayerInventory, Component pTitle) {
         super(pMenu, pPlayerInventory, pTitle);
+        this.imageHeight = 133;
+        this.inventoryLabelY = this.imageHeight - 94;
+    }
+
+    @Override
+    public void render(@NotNull PoseStack pPoseStack, int pMouseX, int pMouseY, float pPartialTick) {
+        this.renderBackground(pPoseStack);
+        super.render(pPoseStack, pMouseX, pMouseY, pPartialTick);
+        //this.renderLabels(pPoseStack, pMouseX, pMouseY);
+        this.renderTooltip(pPoseStack, pMouseX, pMouseY);
     }
 
     @Override
@@ -24,12 +34,5 @@ public class HenhouseScreen extends AbstractContainerScreen<HenhouseMenu> {
         RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
         RenderSystem.setShaderTexture(0, BACKGROUND);
         this.blit(pPoseStack, leftPos, topPos, 0, 0, imageWidth, imageHeight);
-    }
-
-    @Override
-    public void render(@NotNull PoseStack pPoseStack, int pMouseX, int pMouseY, float pPartialTick) {
-        renderBackground(pPoseStack);
-        super.render(pPoseStack, pMouseX, pMouseY, pPartialTick);
-        renderLabels(pPoseStack, pMouseX, pMouseY);
     }
 }